(PE-33645) Update to apacheds 1.5.7
This updates from apacheds 1.5.5 to 1.5.7 to support modern versions of
java. Due to the complexity involved with keeping the embedded test DS
working, combined with the limited additional test value the feature
adds, we no longer set up a TLS listener in tests. That means we only
test the main features of the library and not its ability to make a TLS
connection.