(maint) Override java version in tests
The version of directory-server we use, being from 2010, doesn't
understand the concept of java versions with more than one digit. This
causes test failures because at some point during initialization it
needs to parse the java version. It doesn't actually *use* the version,
it's just parsed as a side effect of loading a static SystemUtils class.
So we can just fake it out with a totally bogus but properly formatted
version number.